+++ near-sdk/src/lib.rs
@@ -2,8 +2,8 @@
 extern crate quickcheck;
 
 pub use near_sdk_macros::{
-    callback, callback_vec, ext_contract, init, metadata, near_bindgen, result_serializer,
-    serializer,
+    callback, callback_vec, ext_contract, init, init_once, metadata, near_bindgen,
+    result_serializer, serializer,
 };
 
 pub mod collections;